+++ test/formatters/samplecode/phpcs/in.php
@@ -1,0 +1,5 @@
+<?php
+
+if(true)
+{
+    echo 'Hello World';}